update dev env
