--- src/components/projects/home/Skills.jsx
@@ -1,52 +1,0 @@
-function Skills() {
-
-    return (
-        <>
-            <section className="skills__section">
-                <h1
-                    className="skills__title"
-                >Get to Know my skills</h1>
-                <div className="skils__box">
-                    <h2
-                        className="skills__subtitle"
-                    >Management and Analysis</h2>
-                    <ul className="skills__list">
-                        <li className="skill__item">{ }</li>
-                        <li className="skill__item">{ }</li>
-                        <li className="skill__item">{ }</li>
-                        <li className="skill__item">{ }</li>
-                        <li className="skill__item">{ }</li>
-                        <li className="skill__item">{ }</li>
-                    </ul>
-
-                </div>
-                <div>
-                    <h2
-                        className="skills__subtitle"
-                    >Web Development</h2>
-                    <ul className="skills__icons">
-                        <li className="skill__icon__item"></li>
-                    </ul>
-
-                </div>
-                <div className="achivements__box">
-                    <div>
-                        <div className="pic__banner">
-                            <img className="emo__pic" src="" alt="" />
-                            <h4>+95% SNEX</h4>
-                            <img className="graph__pic" src="" alt="" />
-                        </div>
-                        <p className="achievement__summary">“As the Product Owner in the Banco BCI project, achieved a 50% increase in the number of operational users (reaching 2K) and a 95% SNEX score (Customer Satisfaction and Efficiency) following the successful launch of the MVP into production.”</p>
-                    </div>
-                </div>
-            </section>
-            <section className="content__section">
-                <div>
-
-                </div>
-            </section>
-        </>
-    );
-}
-
-export default Skills;
