+++ src/styles/home/_skills.scss
@@ -1,0 +1,48 @@
+.skills__section {
+  display: flex;
+  flex-direction: column;
+  margin: 5rem;
+  gap: 44px;
+
+  .skills__title {
+    /* Global Tokens/Headings/H-1 */
+    color: var(--color-neutral-100, #f9fafb);
+    font: var(--font-headings-H-1);
+  }
+}
+
+.skills__article {
+  display: flex;
+  align-items: flex-start;
+  gap: 24px;
+  justify-content: space-between;
+
+  color: var(--color-neutral-100, #f9fafb);
+}
+
+.skills__card {
+  display: flex;
+  padding: 20px 24px;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 24px;
+  flex: 1 0 0;
+  gap: 24px;
+
+  color: var(--color-neutral-100, #f9fafb);
+  font: var(--font-Body-B-2);
+  border-radius: 24px;
+}
+
+.skills {
+  font: var(--font-headings-H-2);
+  color: var(--color-neutral-100, #f9fafb);
+}
+
+.skills__list {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 8px;
+  align-self: stretch;
+}
